requirements.txt
Django==3.1.2
djangorestframework==3.11.0
django-filter==2.4.0
psycopg2==2.8.6
drf-yasg==1.17.1

